6.0

+++ Resources/Anime/routes.py
@@ -1,41 +1,42 @@
 from flask import request
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from uuid import uuid4
 from flask.views import MethodView
 from flask_smorest import abort
-from marshmallow import post_dump
 
 from models import AnimeModel
 from schemas import PostSchema, PostSchemaNested
+
 from . import bp
+# post routes
 
-# Anime routes
-
-@bp.route('/<anime_id>')
+@bp.route('/<post_id>')
 class Post(MethodView):
 
   @bp.response(200, PostSchemaNested)
-  def get(self, anime_id):
-    post = AnimeModel.query.get(anime_id)
+  def get(self, post_id):
+    post = AnimeModel.query.get(post_id)
     if post:
-      print(post.author)
       return post 
-    abort(400, message='Invalid POst')
+    abort(400, message='Invalid Post')
 
+  @jwt_required
   @bp.arguments(PostSchema)
-  def put(self, anime_data ,anime_id):
-    post = AnimeModel.query.get(anime_id)
-    if post:
-      post.body = anime_data['body']
-      post.commit()
-      return {'message': 'anime updated'}, 201
+  def put(self, post_data, post_id):
+    post = AnimeModel.query.get(post_id)
+    if post and post.user_id == get_jwt_identity():
+      post.body = post_data['body']
+      post.commit()   
+      return {'message': 'post updated'}, 201
     return {'message': "Invalid Post Id"}, 400
-
-  def delete(self, anime_id):
-    post = AnimeModel.query.get(anime_id)
-    if post:
+    
+  @jwt_required()
+  def delete(self, post_id):
+    post = AnimeModel.query.get(post_id)
+    if post and post.user_id == get_jwt_identity():
       post.delete()
       return {"message": "Post Deleted"}, 202
-    return {'message':"Invalid Post"}, 400
+    return {'message':"Invalid Post or User"}, 400
 
 @bp.route('/')
 class PostList(MethodView):
@@ -44,12 +45,13 @@ class PostList(MethodView):
   def get(self):
     return AnimeModel.query.all()
   
+  @jwt_required()
   @bp.arguments(PostSchema)
-  def post(self, anime_data):
+  def post(self, post_data):
     try:
       post = AnimeModel()
-      post.user_id = anime_data['user_id']
-      post.body = anime_data['body']
+      post.user_id = get_jwt_identity() 
+      post.body = post_data['body']
       post.commit()
       return { 'message': "Post Created" }, 201
     except:
